pyproject.toml
[tool.poetry]
name = "llm-agent-work"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "llm_agent_work"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.14.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
